+++ generate-xml/generate.nix
@@ -37,6 +37,7 @@ rec
 
   subelem = etype: attrs: contents: sub etype (many (elem etype attrs contents));
 
+  typeConstant = c: x: c;
   typeString = typeConvert "string" builtins.isString id;
   typeInt = typeConvert "int" builtins.isInt builtins.toString;
   typeBoolYesNo = typeConvert "bool" builtins.isBool yesno;